Handle paths with spaces
Some quoting in server_loop that should fix this.  (And preventive
quoting in the command scripts in case the basedir changes.)

So far I've spot-tested these changes with some sample files/directories
and also checked against a regression in my current server setup.
Haven't yet tested a real server setup that uses paths-with-spaces.