Prevent starting of redundant servers
Use "--startas" rather than "--exec" in the start-stop-daemon args for
starting a server.

"--exec" checks for an existing process that matches the executable, but
in our case the executable is going to be shown as bash (rather than the
bash script we ran).  So the match would never actually match.  I
suppose I could check for "bash" instead but that seems generic enough
to not be much of a check.

"--startas" does what we want without looking for an executable match.

We're still checking the PID file of course.