Don't make server_loop find the launch script
Remove the need for server_loop to find a server launch script.  Instead, have the command script tell it everything.

This is not an earthshaking change but I prefer this arrangement for a few reasons.  It's nicer just to have two levels of scripts, with the command script as "master", rather than three levels.  In the three-level setup it wasn't explicit (from looking at the command script) what server launch script would be used, or what variables were for server_loop vs. for the launch script... it just seemed inelegant.

In general this form of the code has less "magic" going on.

I also moved myip into scriptlib in the process of this change.

There's still quite a bit of common code between various command scripts.  Might be good to factor that out and dump it into scriptlib.  On the other hand, having every relevant configuration in the command script makes things simpler to understand.