improve webmin plugin to show server's running status
Show whether a game server is running or not. Also use that info to determine which start/stop buttons to offer for that service.

The server status is loaded at page load time and not auto-refreshed.

Also a few changes here to try to make the default theme of the newer Webmin version play more nicely with my column layout.