indicate when a running server is manually started and not a service
I think in older Debian setups, if you manually started a gameserver script then the service status would just think OK sure yeah I guess I am running. With latest Debian however that is not the case. So if a server is running manually, don't offer any service start/stop actions for it on the webmin UI.

Also as a slight tweak, the Restart action button will be disabled if the Stop button is disabled.